backend/requirements.txt
fastapi[all]>=0.115.6
openai>=1.58.1
pyjwt>=2.10.1
python-docx>=1.1.2
python-dotenv>=1.0.1
python-multipart>=0.0.20
requests>=2.32.3
uvicorn>=0.34.0
websockets>=14.1